pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-streaming-pipeline"
version = "0.1.0"
description = "Kafka ingestion pipeline — API channel producer"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka[schema-registry]>=2.3",
    "httpx>=0.27",
    "fastavro>=1.9",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.3",
    "testcontainers[kafka]>=4.5",
    "hypothesis>=6.100",
    "ruff>=0.4",
]
processing = [
    "apache-flink>=1.20",
    "confluent-kafka[schema-registry]>=2.3",
    "fastavro>=1.9",
    "geoip2>=4.8",
    "prometheus-client>=0.20",
    "opentelemetry-sdk>=1.24",
    "pyyaml",
    "testcontainers[kafka]>=4",
    "hypothesis>=6",
]
scoring = [
    "apache-flink>=1.20",
    "pydantic>=2.0",
    "pyyaml",
    "psycopg2-binary",
    "confluent-kafka[schema-registry]>=2.3",
    "fastavro>=1.9",
    "prometheus-client>=0.20",
    "opentelemetry-sdk>=1.24",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipelines*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests requiring Docker",
    "slow: marks tests as slow (high-volume latency/throughput assertions)",
    "perf: marks tests as performance benchmarks (opt-in only)",
]

[tool.coverage.run]
source = ["pipelines"]
omit = ["tests/*", "*/migrations/*"]
relative_files = true

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["pipelines"]
